update styles for chat page. now responsive
