+++ src/app/globals.css
@@ -24,7 +24,7 @@
 
 body {
   @apply font-pbr bg-pbr-darkpurple text-white p-4;
-  overflow:hidden ;
+  /* overflow:hidden; */
 }
 
 .center {
