+++ src/components/Chat/ChatInfo.tsx
@@ -1,0 +1,15 @@
+import chats_data from "../../../chat_data"
+
+export default function ChatInfo() {
+    return (
+        <div className="x">
+        xxxxxxxxxxxxxxxxxxxx
+        <div className="more">
+            <h1>
+                moreeeeeeeeeeee
+            </h1>
+        </div>
+        </div>
+        
+    )
+}
